pyproject.toml
[project]
name = "mastro"
version = "0.1.0"
description = "My AI Assistant for Astronomical Research"
readme = "README.md"
authors = [
    { name = "yuliu", email = "yuliu@gzu.edu.cn" }
]
requires-python = ">=3.13"
dependencies = [
    "astropy>=7.2.0",
    "astroquery>=0.4.11",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "typer>=0.25.1",
]

[project.scripts]
ads = "mastro.ads.cli:app"
scihub = "mastro.scihub.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
